+++ src/entity/ingredient.rs
@@ -69,8 +69,7 @@ impl Display for Ingredient
     }
 }
 
-
-type DataFromDb = (Option<u8>, Option<String>, Option<String>, Option<f32>,
+type DataFromDb = (Option<u32>, Option<String>, Option<String>, Option<f32>,
                    Option<u16>, Option<u8>, Option<u8>, Option<String>);
 
 impl Ingredient
